+++ Changelog.md
@@ -9,6 +9,7 @@
 - Pick Starshards while holding blueprints now immediately evokes the placed contraption.
 - Added dialogue explanations about artifacts to Halloween Day 7.
 - Added the zoom feature to the Misc Almanac.
+- Added import and export feature for user save data.
 - Added 6 new Random China events.
 
 ## Modified